requirements.txt
analytics-python==1.0.3
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Padding==0.4
pycrypto==2.6.1
python-dateutil==2.4.2
PyYAML==3.11
requests==2.7.0
six==1.9.0
Werkzeug==0.10.4
